+++ SmartHome.Model/Models/UserInfo.cs
@@ -14,7 +14,7 @@ namespace SmartHome.Model.Models
         #region Primitive Properties
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int UserInfoId { get; set; }
+        public long UserInfoId { get; set; }
         public int AppsUserId { get; set; }
         public string LocalId { get; set; }
         [DataType(DataType.Password)]
